Moved Scenario class to uhc\game.
