+++ src/uhc/Loader.php
@@ -13,7 +13,7 @@ use uhc\command\SpectatorCommand;
 use uhc\command\TpallCommand;
 use uhc\command\UHCCommand;
 use uhc\game\GameHeartbeat;
-use uhc\utils\Scenario;
+use uhc\game\Scenario;
 use function is_array;
 use function is_dir;
 use function mkdir;
